Chunk->nonEmptyChunk, wait for production
There is a break of compatibility here with the use of NonEmptyChunkbut this is not really used internally.

Instead, a transaction delimiter is offered to the queue that produces data whenever we think that the user is done. When we see the token coming back (in the form of a promise) we commit the transaction.