Transactional core (#520)
* working PoC (without the hack)

* wip: rewriting tests

* remove tests

Many tests lost value due to re-engineering of the core with
transactions. This also indicates that they were fragile and
a bad idea. E2E tests survived well so it might make more
sense to expand them with all the modules.

* small refactor

* remove code related to recovery

* Chunk->nonEmptyChunk, wait for production

There is a break of compatibility here with the use of NonEmptyChunkbut this is not really used internally.

Instead, a transaction delimiter is offered to the queue that produces data whenever we think that the user is done. When we see the token coming back (in the form of a promise) we commit the transaction.

* fmt

* use internal zio-kafka

* add debug

* include migration path

* upgrade zio-kafka internal

* improve comments and alerts

* use released version

* remove useless config override

* Update core/src/main/scala/tamer/Tamer.scala

Co-authored-by: Julien Jean Paul Sirocchi <julien.sirocchi@gmail.com>

* Update core/src/main/scala/tamer/Tamer.scala

Co-authored-by: Julien Jean Paul Sirocchi <julien.sirocchi@gmail.com>

* Update core/src/main/scala/tamer/Tamer.scala

Co-authored-by: Julien Jean Paul Sirocchi <julien.sirocchi@gmail.com>

* fmt

Co-authored-by: Julien Jean Paul Sirocchi <julien.sirocchi@gmail.com>