remove lag checks
Turns out that checking for the lag will always miss cases whenever tamer gets killed in the middle of a transactions.
This is because it will generate an indefinite amount of aborted transactions (one per failures in a row).
For this reason checking lags alone becomes excessively ambiguous, for example a lag of 5 might mean everything is ok and we are going to skip a failed transaction, but it might also mean that an old version of tamer had incosistent data and needs to be fast-forwarded.
Transactions should guarantee that, whatever the lag is, there is exactly one message (as in exactly-once delivery) pending for comsuption.
The only caveats are that:
 - Upgrades from old tamer must be done on well behaving topics (lag=1)
 - There is only one replica writing to the state topic